Refactoring: moving excess code from R5RCore to other classes
