+++ java-r5rcore/src/org/ipea/r5r/Utils.java
@@ -1,5 +1,10 @@
 package org.ipea.r5r;
 
+import ch.qos.logback.classic.Level;
+import ch.qos.logback.classic.Logger;
+import ch.qos.logback.classic.LoggerContext;
+import org.slf4j.LoggerFactory;
+
 import java.text.DateFormat;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -14,4 +19,22 @@ public class Utils {
         return (int) ((date.getTime() - reference.getTime()) / 1000L);
     }
 
+    public static void setLogMode(String mode) {
+        LoggerContext loggerContext = (LoggerContext) LoggerFactory.getILoggerFactory();
+
+        Logger logger = loggerContext.getLogger("com.conveyal.r5");
+        logger.setLevel(Level.valueOf(mode));
+
+        logger = loggerContext.getLogger("com.conveyal.osmlib");
+        logger.setLevel(Level.valueOf(mode));
+
+        logger = loggerContext.getLogger("com.conveyal.gtfs");
+        logger.setLevel(Level.valueOf(mode));
+
+        logger = loggerContext.getLogger("com.conveyal.r5.profile.ExecutionTimer");
+        logger.setLevel(Level.valueOf(mode));
+
+        logger = loggerContext.getLogger("org.ipea.r5r.R5RCore");
+        logger.setLevel(Level.valueOf(mode));
+    }
 }
